% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tests.R
\name{breslow_day_test}
\alias{breslow_day_test}
\title{Breslow-Day Test for Homogeneity of Odds Ratios}
\usage{
breslow_day_test(contingency_table, odds_ratio = NA, correct = FALSE)
}
\arguments{
\item{contingency_table}{A 2x2xK contingency table where K is the number of strata.}

\item{odds_ratio}{Optionally, the overall odds ratio to use in the test.
If not provided, the Mantel-Haenszel odds ratio is calculated.}

\item{correct}{Logical indicating whether to apply Tarone's correction.}
}
\value{
A list with the Breslow-Day test statistic, p-value, method used, and data name.
}
\description{
Performs the Breslow-Day test for homogeneity of the odds ratios across different
strata of data. Optionally includes Tarone's correction. This test is useful for
investigating whether the effect of an explanatory variable on the outcome is consistent
across different strata.
}
\references{
Breslow NE, Day NE. Statistical methods in cancer research. Volume I -
The analysis of case-control studies. IARC Sci Publ. 1980;(32):5-338.
}
