% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{dx_balanced_accuracy}
\alias{dx_balanced_accuracy}
\title{Calculate Balanced Accuracy}
\usage{
dx_balanced_accuracy(cm, detail = "full", boot = FALSE, bootreps = 1000)
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link{dx_cm}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{boot}{Logical specifying if confidence intervals should be generated
via bootstrapping.  Note, this can be slow.}

\item{bootreps}{The number of bootstrap replications for calculating confidence intervals.}
}
\value{
Depending on the `detail` parameter, returns a numeric value
        representing the calculated metric or a data frame/tibble with
        detailed diagnostics including confidence intervals and possibly other
        metrics relevant to understanding the metric.
}
\description{
Calculates Balanced Accuracy, which is the average of sensitivity (recall) and specificity.
This metric is particularly useful for imbalanced datasets as it accounts for both the
positive and negative classes equally and doesn't inherently favor the majority class.
}
\details{
Balanced Accuracy mitigates the issue of the regular accuracy metric favoring models
that predict the majority class in an imbalanced dataset. By taking the average of
sensitivity and specificity, it gives a better measure of the overall performance
especially when classes are imbalanced or when costs of different errors vary greatly.

The formula for Balanced Accuracy is:
\deqn{Balanced Accuracy = \frac{Sensitivity + Specificity}{2}}{Balanced Accuracy = (Sensitivity + Specificity) / 2}
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth, threshold = 0.5, poslabel = 1)
simple_balanced_accuracy <- dx_balanced_accuracy(cm, detail = "simple")
detailed_balanced_accuracy <- dx_balanced_accuracy(cm)
print(simple_balanced_accuracy)
print(detailed_balanced_accuracy)
}
\seealso{
\code{\link{dx_cm}} to understand how to create and interact with a 'dx_cm' object.

\code{\link{dx_sensitivity}}, \code{\link{dx_specificity}} for the components of balanced accuracy.
}
\concept{metrics}
