% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_diagnostic_table.R
\name{dx_diagnostic_table}
\alias{dx_diagnostic_table}
\title{Generate a dataframe of diagnostic performance over the range of thresholds specified in the main dx function call.}
\usage{
dx_diagnostic_table(
  dx_obj,
  includeFractions = TRUE,
  includeAUC = TRUE,
  includeAccuracy = TRUE,
  includeSensitivity = TRUE,
  includeSpecificity = TRUE,
  includePPV = TRUE,
  includeNPV = TRUE,
  includeOR = TRUE,
  includeF1 = TRUE
)
}
\arguments{
\item{dx_obj}{An object of class dx}

\item{includeFractions}{Logical for inclusion of the fractions appended to the estimates and CIs}

\item{includeAUC}{Logical to include AUC in data frame}

\item{includeAccuracy}{Logical to include Accuracy in data frame}

\item{includeSensitivity}{Logical to include sensitivity in data frame}

\item{includeSpecificity}{Logical to include specificity in data frame}

\item{includePPV}{Logical to include positive predictive value in data frame}

\item{includeNPV}{Logical to include negative predictive value in data frame}

\item{includeOR}{Logical to include Odds Ratio in data frame}

\item{includeF1}{Logical to include F1 in data frame}
}
\description{
Generate a dataframe of diagnostic performance over the range of thresholds specified in the main dx function call.
}
\examples{
dx_obj <- dx(
  data = dx_heart_failure,
  study_name = "Heart Attack Prediction",
  data_description = "Validation Data",
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  threshold_range = c(.1, .2, .3),
  setthreshold = .3,
  grouping_variables = c("AgeGroup", "Sex", "AgeSex")
)
dx_diagnostic_table(dx_obj, includeOR=FALSE, includeF1=FALSE)
}
