% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_forrest.R
\name{dx_edit_cell}
\alias{dx_edit_cell}
\title{Edit a cell within an object returned from dx_forrest}
\usage{
dx_edit_cell(table, row, col, name = "core-fg", ...)
}
\arguments{
\item{table}{a table returned from dx_forrest}

\item{row}{Numeric vector of rows to edit}

\item{col}{Numeric vector of columns to edit}

\item{name}{Name of table layer to edit}

\item{...}{Parameters passed to editGrob such as gpar or hjust.}
}
\description{
A convenient wrapper used to edit cells of a gtable.
}
