% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{dx_f2}
\alias{dx_f2}
\title{Calculate F2 Score with Confidence Intervals}
\usage{
dx_f2(cm, detail = "full", boot = FALSE, bootreps = 1000)
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link[=dx_cm]{dx_cm()}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{boot}{Logical specifying if confidence intervals should be generated
via bootstrapping.  Note, this can be slow.}

\item{bootreps}{The number of bootstrap replications for calculating confidence intervals.}
}
\value{
Depending on the \code{detail} parameter, returns a single numeric value of
F2 or a data frame with the F2 and its confidence intervals.
}
\description{
Calculates the F2 score from a confusion matrix object with an option to
include bootstrapped confidence intervals. The F2 score weights recall
higher than precision.
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth,
  threshold =
    0.5, poslabel = 1
)
simple_f <- dx_f2(cm, detail = "simple")
detailed_f <- dx_f2(cm)
print(simple_f)
print(detailed_f)
}
\seealso{
\code{\link[=dx_fbeta]{dx_fbeta()}}, \code{\link[=dx_f1]{dx_f1()}} for other F-beta scores.
}
\concept{metrics}
