% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{dx_fbeta}
\alias{dx_fbeta}
\title{Calculate F-beta Score with Confidence Intervals}
\usage{
dx_fbeta(cm, beta = 1, detail = "full", boot = FALSE, bootreps = 1000)
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link{dx_cm}}.}

\item{beta}{The beta value determining the weight of precision in the F-score.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{boot}{Logical specifying if confidence intervals should be generated
via bootstrapping.  Note, this can be slow.}

\item{bootreps}{The number of bootstrap replications for calculating confidence intervals.}
}
\value{
Depending on the `detail` parameter, returns a single numeric value of
        F-beta or a data frame with the F-beta and its confidence intervals.
}
\description{
Calculates the F-beta score from a confusion matrix object with an option
to include bootstrapped confidence intervals. The F-beta score is a
generalization of the F1 score, allowing different importance to precision
and recall via the beta parameter.
}
\seealso{
\code{\link{dx_f1}}, \code{\link{dx_f2}} for specific F-beta scores.
}
\concept{metrics}
