% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tests.R
\name{dx_fishers_exact}
\alias{dx_fishers_exact}
\title{Fisher's Exact Test for Independence in a 2x2 Table}
\usage{
dx_fishers_exact(cm, detail = "full")
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link{dx_cm}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}
}
\value{
Depending on the `detail` parameter:
        - if "simple": a single numeric value representing the p-value of
          Fisher's Exact test.
        - if "full": a data frame with the Fisher's Exact test result, including
          the p-value and method note.
}
\description{
Conducts Fisher's Exact test of independence on a 2x2 confusion matrix derived
from binary classification results, assessing the significance of the association
between the observed and expected frequencies.
}
\details{
Fisher's Exact Test is used to examine the significance of the association
between the variables in a 2x2 contingency table, particularly useful when sample
sizes are small. Unlike the chi-square test, it does not rely on large sample
distribution approximations and is hence exact. It's especially preferred when the
data has small expected frequencies in one or more cells of the table. A low p-value
indicates a significant association between the predicted and actual binary classifications.
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth,
  threshold = 0.3, poslabel = 1
)
simple <- dx_fishers_exact(cm, detail = "simple")
detailed <- dx_fishers_exact(cm)
print(simple)
print(detailed)
}
\seealso{
\code{\link{dx_cm}} for creating a 'dx_cm' object.
}
\concept{tests}
