% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_forest.R
\name{dx_forest}
\alias{dx_forest}
\title{Create table with odds ratios displayed graphically}
\usage{
dx_forest(
  dx_obj,
  breaks = NA,
  limits = NA,
  tick_label_size = 6.5,
  trans = c(NA, "log10"),
  return_grid = FALSE,
  filename = NA
)
}
\arguments{
\item{dx_obj}{An object of class dx}

\item{breaks}{A numeric vector of breaks to include on axis ticks.  If left empty, breaks will be determined using the min and max value from 95\% CIs.}

\item{limits}{Limits for axis ticks.  Ticks will be generates using base::breaks.  Ignored if breaks are passed.}

\item{tick_label_size}{Font size for axis labels.}

\item{trans}{Method to transform the odds ratio by.  Currently, only log10 is supported.}

\item{return_grid}{Should a grid object be returned? If FALSE, grid is drawn using grid.draw.}

\item{filename}{File bane to create on disk.  If left NA, no file will be created.}
}
\description{
Generate a table of diagnostic measures
}
\examples{

dx_obj <- dx(
  data = dx_heart_failure,
  study_name = "Heart Attack Prediction",
  data_description = "Validation Data",
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  threshold_range = c(.1,.2,.3),
  setthreshold = .3,
  grouping_variables = c("AgeGroup", "Sex", "AgeSex")
)
dx_forest(dx_obj)
dx_forest(dx_obj, trans = "log10")
}
