% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_forrest.R
\name{dx_forrest}
\alias{dx_forrest}
\title{Create table with odds ratios displayed graphically}
\usage{
dx_forrest(dx_obj, breaks = NA, limits = NA, tick_label_size = 6.5,
  return_grid = FALSE)
}
\arguments{
\item{dx_obj}{An object of class dx}

\item{breaks}{A numeric vector of breaks to include on axis ticks.  If left empty, breaks will be determined using the min and max value from 95\% CIs.}

\item{limits}{Limits for axis ticks.  Ticks will be generates using base::breaks.  Ignored if breaks are passed.}

\item{tick_label_size}{Font size for axis labels.}

\item{return_grid}{Should a grid object be returned? If FALSE, grid is drawn using grid.draw.}
}
\description{
Generate a table of diagnostic measures
}
\examples{
dx_obj <- dx(
  data = dx_heart_failure,
  study_name = "Heart Attack Prediction",
  data_description = "Validation Data",
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  threshold_range = c(.1,.2,.3),
  setthreshold = .3,
  grouping_variables = c("AgeGroup", "Sex", "AgeSex")
)
dx_forrest(dx_obj)
}
