% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{dx_fowlkes_mallows}
\alias{dx_fowlkes_mallows}
\title{Calculate Fowlkes-Mallows Index}
\usage{
dx_fowlkes_mallows(cm, detail = "full", boot = FALSE, bootreps = 1000)
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link{dx_cm}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{boot}{Logical specifying if confidence intervals should be generated
via bootstrapping.  Note, this can be slow.}

\item{bootreps}{The number of bootstrap replications for calculating confidence intervals.}
}
\value{
Depending on the `detail` parameter, returns a numeric value
        representing the calculated metric or a data frame/tibble with
        detailed diagnostics including confidence intervals and possibly other
        metrics relevant to understanding the metric.
}
\description{
Calculates the Fowlkes-Mallows Index (FM Index) for the provided confusion matrix.
FM Index is the geometric mean of precision and recall, providing a balance
measure between these two metrics.
}
\details{
Fowlkes-Mallows Index is defined as the geometric mean of the precision (Positive Predictive Value)
and recall (True Positive Rate or Sensitivity). It's a useful measure when you want a balance between
precision and recall without the harshness of the harmonic mean used in F1 score. A higher Fowlkes-Mallows
Index indicates better precision and recall balance.

The formula for Fowlkes-Mallows Index is:
\deqn{FM = \sqrt{Precision \times Recall}}{FM = sqrt(Precision * Recall)}
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth, threshold = 0.5, poslabel = 1)
simple_fm_index <- dx_fowlkes_mallows(cm, detail = "simple")
detailed_fm_index <- dx_fowlkes_mallows(cm)
print(simple_fm_index)
print(detailed_fm_index)
}
\seealso{
\code{\link{dx_cm}} to understand how to create and interact with a 'dx_cm' object.

\code{\link{dx_ppv}}, \code{\link{dx_sensitivity}} for components of FM Index.
}
\concept{metrics}
