% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{dx_g_mean}
\alias{dx_g_mean}
\title{Calculate G-mean}
\usage{
dx_g_mean(cm, detail = "full", boot = FALSE, bootreps = 1000)
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link{dx_cm}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{boot}{Logical specifying if confidence intervals should be generated
via bootstrapping.  Note, this can be slow.}

\item{bootreps}{The number of bootstrap replications for calculating confidence intervals.}
}
\value{
Depending on the `detail` parameter, returns a numeric value
        representing the calculated metric or a data frame/tibble with
        detailed diagnostics including confidence intervals and possibly other
        metrics relevant to understanding the metric.
}
\description{
Calculates the Geometric Mean (G-mean) for the provided confusion matrix.
G-mean is a measure of a model's performance that considers both the sensitivity
(True Positive Rate) and specificity (True Negative Rate), especially useful in
imbalanced datasets.
}
\details{
G-mean is the geometric mean of sensitivity and specificity. It tries to maximize
the accuracy on each of the two classes while keeping these accuracies balanced.
For a classifier to achieve a high G-mean score, it must perform well on both
positive and negative classes.

The formula for G-mean is:
\deqn{G-mean = \sqrt{Sensitivity \times Specificity}}{G-mean = sqrt(Sensitivity * Specificity)}
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth, threshold = 0.5, poslabel = 1)
simple_g_mean <- dx_g_mean(cm, detail = "simple")
detailed_g_mean <- dx_g_mean(cm)
print(simple_g_mean)
print(detailed_g_mean)
}
\seealso{
\code{\link{dx_cm}} to understand how to create and interact with a 'dx_cm' object.

\code{\link{dx_sensitivity}}, \code{\link{dx_specificity}} for components of G-mean.
}
\concept{metrics}
