% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_summary.R
\name{dx_grouped_summary}
\alias{dx_grouped_summary}
\title{Grouped Summary}
\usage{
dx_grouped_summary(data, group_varname, pred_varname, threshold,
  true_varname, poslabel, citype, bootreps, bootseed, doboot)
}
\description{
Grouped Summary
}
