% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_heart_failure.R
\docType{data}
\name{dx_heart_failure}
\alias{dx_heart_failure}
\title{Heart attack outcomes and predictions}
\format{
A data frame with 261 rows and 5 variables:
\describe{
  \item{AgeGroup}{Age group}
  \item{Sex}{sex}
  \item{truth}{Heart failure (outcome)}
  \item{predicted}{Predicted outcome from model}
  \item{AgeSex}{Age and sex group}
}
}
\source{
\url{https://www.kaggle.com/imnikhilanand/heart-attack-prediction}
}
\usage{
dx_heart_failure
}
\description{
A data set containing heart attack outcomes for 261 patients along with
predicted probabilities of having a heart attack.
}
\keyword{datasets}
