% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{dx_lrt_pos}
\alias{dx_lrt_pos}
\title{Calculate Positive Likelihood Ratio}
\usage{
dx_lrt_pos(cm, detail = "full", ...)
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link{dx_cm}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{...}{Additional arguments to pass to metric_binomial function, such as
`citype` for type of confidence interval method.}
}
\description{
Calculates the Positive Likelihood Ratio (LR+) from a confusion matrix object.
LR+ compares the probability of a positive test result among patients with the
disease to the probability of a positive test result among patients without the disease.
}
\details{
The positive likelihood ratio is calculated as (TP / (TP + FN)) / (FP / (FP + TN)).
It is used to assess the diagnostic usefulness of a test. A LR+ much greater than 1
indicates a good diagnostic test that can confidently confirm the disease when the
test is positive.
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth,
  threshold =
    0.5, poslabel = 1
)
simple_lrp <- dx_lrt_pos(cm, detail = "simple")
detailed_lrp <- dx_lrt_pos(cm)
print(simple_lrp)
print(detailed_lrp)
}
\concept{metrics}
