% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_plots.R
\name{dx_plot_cm}
\alias{dx_plot_cm}
\title{Plot Confusion Matrix with Metrics}
\usage{
dx_plot_cm(dx_obj, palette = c("#e5eef7", "#0057B8"), levels = c("-", "+"))
}
\arguments{
\item{dx_obj}{An object of class "dx" containing the diagnostic measurements.}

\item{palette}{A character vector of length 2 specifying the colors for the
low and high ends of the fill gradient used in the plot. Defaults to
c("#e5eef7", "#0057B8").}

\item{levels}{A character vector of length 2 specifying the labels for negative
and positive classes, respectively. Defaults to c("-", "+").}
}
\value{
A \code{ggplot} object that represents the confusion matrix with additional
performance metrics.
}
\description{
Creates a graphical representation of a confusion matrix from a \code{dx} object.
The plot displays the counts of true positives, false negatives, true negatives,
and false positives. Additionally, it annotates the plot with relevant performance
metrics such as Sensitivity, Specificity, PPV, and NPV, along with their
confidence intervals where applicable.
}
\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dx_obj <- dx(
  data = dx_heart_failure,
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  setthreshold = .3
)
dx_plot_cm(dx_obj)
\dontshow{\}) # examplesIf}
}
