% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_plots.R
\name{dx_plot_probabilities}
\alias{dx_plot_probabilities}
\title{Plot Predicted Probabilities}
\usage{
dx_plot_probabilities(
  dx_obj,
  plot_type = "histogram",
  bins = NULL,
  fill_color = "#0057B8"
)
}
\arguments{
\item{dx_obj}{A \code{dx} object}

\item{plot_type}{Character string specifying the type of plot: "histogram" or "density".}

\item{bins}{Optional; number of bins for the histogram (relevant if plot_type = "histogram").}

\item{fill_color}{Color to fill the bars or density plot.}
}
\value{
A ggplot object representing the distribution of predicted probabilities.
}
\description{
Generates a plot showing the distribution of predicted probabilities.
Offers options to show as a histogram or density plot.
}
\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dx_obj <- dx(
  data = dx_heart_failure,
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  setthreshold = .3
)
dx_plot_probabilities(dx_obj, plot_type = "histogram", bins = 30)
dx_plot_probabilities(dx_obj, plot_type = "density")
\dontshow{\}) # examplesIf}
}
