% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_plots.R
\name{dx_plot_thresholds}
\alias{dx_plot_thresholds}
\title{Plot Diagnostic Measures across Thresholds}
\usage{
dx_plot_thresholds(dx_obj)
}
\arguments{
\item{dx_obj}{A \code{dx} object containing threshold-based statistics, including values
for various diagnostic measures at different thresholds.}
}
\value{
A \code{ggplot} object representing the diagnostic measures across thresholds
with the ability to further customize if desired.
}
\description{
Generates a line plot to visualize various diagnostic measures across different
threshold values for a binary classification model. This visualization can help in
selecting an optimal threshold based on the trade-offs between different measures.
}
\details{
The function plots multiple lines representing different diagnostic measures such
as NPV, PPV, sensitivity, specificity, and F1 score across a range of threshold
values. Each line corresponds to a specific metric, illustrating how the measure
changes as the classification threshold is varied. A vertical dashed line indicates
the set threshold in the \code{dx} object for reference.

This plot is particularly useful for understanding the behavior of a classifier
under different operating conditions and for identifying a threshold that balances
the trade-offs between various measures according to the specific needs of the
application.
}
\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dx_obj <- dx(
  data = dx_heart_failure,
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  setthreshold = .3
)
dx_plot_thresholds(dx_obj)
\dontshow{\}) # examplesIf}
}
