% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{dx_ppv_prevalence}
\alias{dx_ppv_prevalence}
\title{Calculate Positive Predictive Value (PPV) at Target Prevalence}
\usage{
dx_ppv_prevalence(cm, prevalence, detail = "full", ...)
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link[=dx_cm]{dx_cm()}}.}

\item{prevalence}{Numeric value between 0 and 1, representing the target prevalence
for adjusting the PPV calculation.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{...}{Additional arguments to pass to metric_binomial function, such as
\code{citype} for type of confidence interval method.}
}
\value{
Depending on the \code{detail} parameter, returns a numeric value
representing the calculated metric or a data frame/tibble with
detailed diagnostics including confidence intervals and possibly other
metrics relevant to understanding the metric.
}
\description{
Computes the Positive Predictive Value (PPV) adjusted to a specified prevalence
level. This function is useful for understanding classifier performance in settings
where the actual prevalence of the condition may differ from that in the initial data.
}
\details{
This function calculates the PPV using the formula:
\deqn{PPV = \frac{Sensitivity \times Prevalence}{(Sensitivity \times Prevalence) + (1 - Specificity) \times (1 - Prevalence)}}
where \code{Sensitivity} is the true positive rate and \code{Specificity} is the true negative rate.
Adjusting PPV for prevalence allows a more accurate assessment of the classifier’s performance
in different population settings.

Confidence intervales are calucated using a simple logit (Mercaldo et al, 2007)
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth,
  threshold = 0.5, poslabel = 1
)
# Calculate PPV at a prevalence of 0.1
dx_ppv_prevalence(cm, prevalence = 0.1)

}
\seealso{
\code{\link[=dx_cm]{dx_cm()}}, \code{\link[=dx_ppv]{dx_ppv()}} for the non-prevalence adjusted PPV.
}
\concept{metrics}
