% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{dx_prevalence}
\alias{dx_prevalence}
\title{Calculate Prevalence}
\usage{
dx_prevalence(cm, detail = "full", ...)
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link{dx_cm}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{...}{Additional arguments to pass to metric_binomial function, such as
`citype` for type of confidence interval method.}
}
\description{
Calculates Prevalence, which is the proportion of cases that are positive for
the condition of interest over the total number of cases. Prevalence provides
a measure of how widespread a condition is within the population at a given time.
}
\details{
Prevalence is a measure of the burden of a condition or disease in a population.
It is an important measure in epidemiology and health service planning as it helps
to understand the level of disease in a population at a given time. Unlike other
metrics that are based on the classifier's performance, prevalence is a measure of
the actual condition being tested.

The formula for Prevalence is:
\deqn{Prevalence = \frac{Number of Current Cases (Positives)}{Total Number of Cases}}{Prevalence = Pos / N}
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth, threshold =
            0.5, poslabel = 1)
simple_prevalence <- dx_prevalence(cm, detail = "simple")
detailed_prevalence <- dx_prevalence(cm)
print(simple_prevalence)
print(detailed_prevalence)
}
\seealso{
\code{\link{dx_cm}} to understand how to create and interact with a
         'dx_cm' object.
}
\concept{metrics}
