% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_report.R
\name{dx_report}
\alias{dx_report}
\title{Render diagnostic report}
\usage{
dx_report(
  dx_obj,
  roc_options = list(),
  forest_options = list(),
  output_dir = getwd(),
  output_file = "test.html"
)
}
\arguments{
\item{dx_obj}{An object of class dx}

\item{roc_options}{Arguments to be passed to dx_rox}

\item{forest_options}{Arguments to be passed to dx_forest}

\item{output_dir}{Directory to output report into}

\item{output_file}{Filename of generated report}
}
\description{
Render diagnostic report
}
\examples{
\dontrun{
dx_obj <- dx(
  data = dx_heart_failure,
  study_name = "Heart Attack Prediction",
  data_description = "Validation Data",
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  threshold_range = c(.1, .2, .3),
  setthreshold = .3,
  grouping_variables = c("AgeGroup", "Sex", "AgeSex")
)
dx_report(dx_obj, roc_options = list(curve_color = "blue"))
}
}
