% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_set_options.R
\name{dx_set_options}
\alias{dx_set_options}
\title{Set options for diagnostic analysis}
\usage{
dx_set_options(data, study_name, data_description,
  classlabels = c("Negative", "Positive"), threshold_range = NA,
  outcome_label, pred_varname, true_varname, setthreshold = 0.5,
  poslabel = 1, grouping_variables = NA, citype = "exact",
  bootreps = 2000, bootseed = 20191015, doboot = FALSE,
  roc_curve_color = "red", roc_text_color = "black",
  roc_add_text = TRUE, roc_add_ref_lines = TRUE,
  roc_add_fractions = TRUE, roc_summary_stats = c(1, 2, 3, 4, 5, 6, 7,
  8), roc_filename = paste0(study_name, "_ROC", data_description,
  ".pdf"))
}
\arguments{
\item{data}{A tbl.}

\item{study_name}{Name of study to be used in the report (string)}

\item{data_description}{Description of data to be used in the report (string)}

\item{classlabels}{Labels for predicted variable.  Needs to be 0, 1 order.}

\item{threshold_range}{Optional. A numeric vector of thresholds to loop over.}

\item{outcome_label}{Label for outcome (string)}

\item{pred_varname}{Colomn name containing AI prediction (string)}

\item{true_varname}{Column name containt AI reference standard (string)}

\item{setthreshold}{A numeric value represnting the threshold used to identify AI prediction}

\item{poslabel}{Positive class.  Variable should be coded as 0/1 with 1 being the event}

\item{grouping_variables}{Character vector of variable names to be summarized by.
These variables should all be factors.}

\item{bootreps}{Number of bootstrap samples used to generate F1 score CI}

\item{bootseed}{Seed value to be used when calculating bootsraped CI's}

\item{doboot}{Logical.  Generate bootstrap estimate of F1 confidence interval?}

\item{roc_curve_color}{Color of ROC curve}

\item{roc_text_color}{Color of text on ROC curve}

\item{roc_add_text}{(Logical) Include stats on ROC curve?}

\item{roc_add_ref_lines}{(Logical) Include reference lines on ROC curve?}

\item{roc_add_fractions}{(Logical) Include fractions on ROC curve?}

\item{roc_summary_stats}{Statistics to include on ROC curve}

\item{roc_filename}{Name of file output for ROC pdf file}

\item{citpye}{Confidence interval type.}
}
\description{
Use this function to return a list of tuning parameters to analyze your
diagnostic test
}
