% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_summary.R
\name{dx_summary}
\alias{dx_summary}
\title{Summarize diagnostic results for a model}
\usage{
dx_summary(data, pred_varname, threshold, true_varname, poslabel = 1,
  citype = "exact", bootreps = 2000, bootseed = 20190617,
  doboot = TRUE)
}
\description{
Summarize diagnostic results for a model
}
