% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tests.R
\name{expected_cell_count}
\alias{expected_cell_count}
\title{Calculate Expected Cell Count}
\usage{
expected_cell_count(a_value, marginals, odds_ratio)
}
\arguments{
\item{a_value}{The observed count for cell 'a' in a single stratum.}

\item{marginals}{A list containing the 'row' and 'col' marginal sums for the stratum.}

\item{odds_ratio}{The overall odds ratio to use in the calculation.}
}
\value{
The expected count for cell 'a'.
}
\description{
Calculates the expected count of the top left cell (a) under the null hypothesis
of homogeneity of odds ratios across strata.
}
\keyword{internal}
