% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_roc.R
\name{get_roc}
\alias{get_roc}
\title{Return an pROC::roc object for a dx object}
\usage{
get_roc(true_varname, pred_varname, data, direction)
}
\arguments{
\item{true_varname}{Column name containing AI reference standard (string)}

\item{pred_varname}{Column name containing AI prediction (string)}

\item{data}{A tbl.}

\item{direction}{Direction for roc comparison.  See ?pROC::roc}
}
\description{
Return an pROC::roc object for a dx object
}
