% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{informedness}
\alias{informedness}
\alias{dx_informedness}
\alias{dx_youden_j}
\title{Calculate Informedness}
\usage{
dx_informedness(cm, detail = "full", boot = FALSE, bootreps = 1000)

dx_youden_j(cm, detail = "full", boot = FALSE, bootreps = 1000)
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link{dx_cm}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{boot}{Logical specifying if confidence intervals should be generated
via bootstrapping.  Note, this can be slow.}

\item{bootreps}{The number of bootstrap replications for calculating confidence intervals.}
}
\description{
Calculates Informedness for the provided confusion matrix. Informedness is a combined measure
of Sensitivity (True Positive Rate) and Specificity (True Negative Rate). It reflects the
probability that a classifier is informed about the true class, ranging from -1 to 1.
}
\details{
Informedness is defined as \code{Informedness = Sensitivity + Specificity - 1}. It is the sum of the true positive rate
and the true negative rate minus one. It's a useful measure when you want to consider both
the sensitivity and specificity of a test. A higher informedness indicates better overall performance
of the classifier in distinguishing between the classes.

The formula for Informedness is:
\deqn{Informedness = Sensitivity + Specificity - 1}{Informedness = Sensitivity + Specificity - 1}
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth, threshold = 0.5, poslabel = 1)
simple_informedness <- dx_informedness(cm)
detailed_informedness <- dx_informedness(cm, detail = "full")
print(simple_informedness)
print(detailed_informedness)
}
\seealso{
\code{\link{dx_cm}} to understand how to create and interact with a 'dx_cm' object.

\code{\link{dx_sensitivity}}, \code{\link{dx_specificity}} for components of informedness.
}
\concept{metrics}
