% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{specificity}
\alias{specificity}
\alias{dx_specificity}
\alias{dx_tnr}
\title{Calculate Specificity (True Negative Rate)}
\usage{
dx_specificity(cm, detail = "full", ...)

dx_tnr(cm, detail = "full", ...)
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link{dx_cm}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{...}{Additional arguments to pass to metric_binomial function, such as
`citype` for type of confidence interval method.}
}
\description{
Calculates Specificity, also known as the True Negative Rate (TNR), which is the
proportion of actual negatives that are correctly identified as such by the classifier.
Specificity is a key measure in evaluating the effectiveness of a classifier in
identifying negative instances.
}
\details{
Specificity or TNR measures how well the classifier can identify negative instances,
which is critical in situations where false positives carry a high cost. A higher
specificity indicates a better performance in recognizing negative instances and
avoiding false alarms.

The formula for Specificity is:
\deqn{Specificity = \frac{True Negatives}{True Negatives + False Positives}}{Specificity = TN / (TN + FP)}
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth, threshold =
            0.5, poslabel = 1)
simple_specificity <- dx_specificity(cm)
detailed_specificity <- dx_specificity(cm, detail = "full")
print(simple_specificity)
print(detailed_specificity)
}
\seealso{
\code{\link{dx_cm}} to understand how to create and interact with a
         'dx_cm' object.
}
\concept{metrics}
