% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tests.R
\name{variance_a}
\alias{variance_a}
\title{Calculate Variance of Cell Count}
\usage{
variance_a(tilde_a_value, marginals)
}
\arguments{
\item{tilde_a_value}{The expected count for cell 'a' in a single stratum.}

\item{marginals}{A list containing the 'row' and 'col' marginal sums for the stratum.}
}
\value{
The variance of the count for cell 'a'.
}
\description{
Calculates the variance of the top left cell (a) count under the null hypothesis
of homogeneity of odds ratios across strata.
}
\keyword{internal}
