% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_cm.R
\name{dx_cm}
\alias{dx_cm}
\title{Plot ROC curve for a dx object}
\usage{
dx_cm(dx_obj, palette = c("#FFFFFF", "#0057B8"), levels = c("-", "+"))
}
\arguments{
\item{dx_obj}{An object of class dx}

\item{palette}{A character vector of length 2.  Colors for low and high ends of the fill gradient}

\item{levels}{A character vector of length 2.  x an y labels for +/- classes}
}
\description{
Plot ROC curve for a dx object
}
\examples{
dx_obj <- dx(
  data = dx_heart_failure,
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  threshold_range = c(.1, .2, .3),
  setthreshold = .3,
  grouping_variables = c("AgeGroup", "Sex", "AgeSex")
)
dx_cm(dx_obj)
}
