% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_diagnostic_table.R
\name{dx_diagnostic_table}
\alias{dx_diagnostic_table}
\title{Generate a dataframe of diagnostic performance over the range of thresholds specified in the main dx function call.}
\usage{
dx_diagnostic_table(
  dx_obj,
  measures = c("AUC", "Accuracy", "Sensitivity", "Specificity",
    "Positive Predictive Value", "Negative Predictive Value", "Odds Ratio", "F1 Score"),
  fraction = TRUE
)
}
\arguments{
\item{dx_obj}{An object of class dx}

\item{measures}{Vector of diagnostic measures to include}

\item{fraction}{Logical. Should fractions be included for measures (where applicable)?}
}
\description{
Generate a dataframe of diagnostic performance over the range of thresholds specified in the main dx function call.
}
\examples{
dx_obj <- dx(
  data = dx_heart_failure,
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  threshold_range = c(.1, .2, .3),
  setthreshold = .3,
  grouping_variables = c("AgeGroup", "Sex", "AgeSex")
)
dx_diagnostic_table(dx_obj)
}
