% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_forest.R
\name{dx_forest}
\alias{dx_forest}
\title{Create table with odds ratios displayed graphically}
\usage{
dx_forest(
  dx_obj,
  fraction = FALSE,
  breaks = NA,
  limits = NA,
  tick_label_size = 6.5,
  trans = c(NA, "log10"),
  measures = c("AUC", "Sensitivity", "Specificity", "Odds Ratio"),
  return = c("ggplot", "grid"),
  filename = NA,
  header_bg = "white",
  header_col = "black",
  body_bg = c("#e6e4e2", "#ffffff"),
  footer_bg = "#b8b6b4",
  footer_col = "black",
  header_fontsize = 10,
  body_fontsize = 9,
  fraction_multiline = FALSE,
  or_lwd = 0.8,
  or_size = 0.35,
  body_or_col = "black",
  footer_or_col = footer_col
)
}
\arguments{
\item{dx_obj}{An object of class dx}

\item{fraction}{Logical.  Should the fraction for sensitivity
and specificity be shown?}

\item{breaks}{A numeric vector of breaks to include on axis ticks.
If left empty, breaks will be determined using
the min and max value from 95\% CIs.}

\item{limits}{Limits for axis ticks.
Ticks will be generates using base::breaks.
Ignored if breaks are passed.}

\item{tick_label_size}{Font size for axis labels.}

\item{trans}{Method to transform the odds ratio by.
Currently, only log10 is supported.}

\item{measures}{Measures to be included in the plot}

\item{return}{Should a grid or ggplot object be returned?}

\item{filename}{File bane to create on disk.
If left NA, no file will be created.}

\item{header_bg}{Background color of the header}

\item{header_col}{Color of text in the header}

\item{body_bg}{Background color of table rows.  If values are less than total
number of rows, values are repeated.}

\item{footer_bg}{Background color if the footer row.}

\item{footer_col}{Color of the footer row.}

\item{header_fontsize}{Font size of header text}

\item{body_fontsize}{Font size of body text}

\item{fraction_multiline}{Logical. Should fractions be split onto 2 lines?}

\item{or_lwd}{Line width for OR}

\item{or_size}{Size of OR point}

\item{body_or_col}{Color of odds ratios in the table body}

\item{footer_or_col}{Color of odds ratios in the table footer}
}
\description{
Generate a table of diagnostic measures
}
\examples{

dx_obj <- dx(
  data = dx_heart_failure,
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  threshold_range = c(.1, .2, .3),
  setthreshold = .3,
  grouping_variables = c("AgeGroup", "Sex", "AgeSex")
)
dx_forest(dx_obj)
dx_forest(dx_obj, trans = "log10")
}
