% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{dx_nir}
\alias{dx_nir}
\title{Calculate No Information Rate (NIR)}
\usage{
dx_nir(cm, detail = "full")
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link[=dx_cm]{dx_cm()}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}
}
\value{
Depending on the \code{detail} parameter, returns a numeric value
representing the calculated metric or a data frame/tibble with
detailed diagnostics including confidence intervals and possibly other
metrics relevant to understanding the metric.
}
\description{
The No Information Rate is the proportion of the largest class in the actual outcomes.
It represents the accuracy that a naive model would achieve by always predicting
the most frequent class. It's a baseline measure for classification performance.
}
\examples{
cm <- dx_cm(dx_heart_failure$predicted, dx_heart_failure$truth, threshold = 0.5, poslabel = 1)
nir <- dx_nir(cm)
print(nir)
}
\concept{metrics}
