% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_roc.R
\name{dx_roc}
\alias{dx_roc}
\title{Plot ROC curve for a dx object}
\usage{
dx_roc(
  dx_obj,
  curve_color = "red",
  text_color = "black",
  add_text = TRUE,
  add_ref_lines = TRUE,
  add_fractions = TRUE,
  summary_stats = c(1, 2, 3, 4, 5, 6, 7, 8),
  filename = NA
)
}
\arguments{
\item{dx_obj}{An object of class dx}

\item{curve_color}{Color of ROC curve}

\item{text_color}{Color of text on ROC curve}

\item{add_text}{(Logical) Include stats on ROC curve?}

\item{add_ref_lines}{(Logical) Include reference lines on ROC curve?}

\item{add_fractions}{(Logical) Include fractions on ROC curve?}

\item{summary_stats}{Statistics to include on ROC curve}

\item{filename}{File bane to create on disk.  If left NA, no file will be created.}
}
\description{
Plot ROC curve for a dx object
}
\examples{
dx_obj <- dx(
  data = dx_heart_failure,
  study_name = "Heart Attack Prediction",
  data_description = "Validation Data",
  true_varname = "truth",
  pred_varname = "predicted",
  outcome_label = "Heart Attack",
  threshold_range = c(.1,.2,.3),
  setthreshold = .3,
  grouping_variables = c("AgeGroup", "Sex", "AgeSex")
)
dx_roc(dx_obj)
}
