% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_metrics.R
\name{metrics-params}
\alias{metrics-params}
\alias{metricparams}
\title{Shared Parameters for Diagnostic Metrics}
\usage{
metricparams()
}
\arguments{
\item{cm}{A dx_cm object created by \code{\link{dx_cm}}.}

\item{detail}{Character specifying the level of detail in the output:
"simple" for raw estimate, "full" for detailed estimate
including 95\% confidence intervals.}

\item{...}{Additional arguments to pass to metric_binomial function, such as
`citype` for type of confidence interval method.}

\item{boot}{Logical specifying if confidence intervals should be generated
via bootstrapping.  Note, this can be slow.}

\item{bootreps}{The number of bootstrap replications for calculating confidence intervals.}
}
\value{
Depending on the `detail` parameter, returns a numeric value
        representing the calculated metric or a data frame/tibble with
        detailed diagnostics including confidence intervals and possibly other
        metrics relevant to understanding the metric.
}
\description{
Shared Parameters for Diagnostic Metrics
}
\keyword{internal}
